Separate serialize/deserialize methods into own helper
